requirements.txt
flask
flask-login
flask-sqlalchemy
sentence-transformers
faiss-cpu
pypdf
python-docx
openai
werkzeug
gunicorn